// src/app/shared/models/task/task-param.model.ts
export class TaskParamModel {
  taskId: string;
  customerName: string;
  contactNumber: string;
  customerEmail: string;
  orderId: string;
  wareHouseId: string;
  dropLatitude: string;
  dropLongitude: string;
  amount: string;
  paymentType: string;
  deliveryDate: string;
  assignDrivers: string[];
  taskStatus: string;
  createdBy: string;
  modifiedBy: string;
  createdOn: string;
  modifiedOn: string;
  dropAddress: string;
  baseFare: string;
  totalFare: string;
  vehicleType: string;
}